refactor(lift): factor shared lifter core into reusable lib (#219)
Both kits' lift adapters and LSP plugins now share a single source of
truth for the IR-emission logic. The binaries are thin shells around
the core; the cores live alongside the existing lift libraries.

Go (real duplication eliminated):
- provekit-lift-go-validator gains public LiftValidateTags / LiftValidateTag
  / GoSortFromTypeName entry points (sort-based, source-agnostic core).
- LiftStruct (reflection-driven, used by the batch CLI) delegates to
  LiftValidateTags as before; behavior unchanged.
- cmd/provekit-lsp-go imports validator and drops its 100-line duplicated
  goKind-string-based reimplementation. liftStructFromAST now derives
  ir.Sort from the AST type expression and calls into the shared core,
  so both binaries emit byte-identical IR for the same tag string.

C# (already partly factored; symmetric core added):
- Provekit.Lift.Core (new) holds the shared annotation scanner
  (//provekit:contract / //provekit:implement) and the Roslyn-driven
  source-to-IR pipeline (compile -> DataAnnotations lift -> annotation
  scan).
- Provekit.Lsp.Plugin drops its annotation scanner and Roslyn pipeline,
  delegating to SourceLifter.LiftSource. The Microsoft.CodeAnalysis.CSharp
  package reference moves to Core.
- Future C# batch-CLI lift binaries can drop straight onto SourceLifter.

Verification:
- Go: cmd/provekit-lift-go and cmd/provekit-lsp-go build clean. LSP
  tests pass (7/7). validator tests show the 4 pre-existing
  json.Marshal-key-order failures unchanged from baseline; not in scope
  for this refactor.
- C#: full solution builds; 61/61 tests pass (Provekit.Tests 44,
  Provekit.Lift.Linq.Tests 9, Provekit.Lift.DataAnnotations 8).
- LSP smoke tests for both kits emit byte-identical declarations to
  pre-refactor output.

Scope: go and csharp (the two kits with both lift adapter and LSP plugin
today). Other kits (python, ruby, zig) follow as separate PRs if needed.
