+++ implementations/csharp/Provekit.Lift.Core/Provekit.Lift.Core.csproj
@@ -1,17 +1,19 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <RootNamespace>Provekit.Lsp.Plugin</RootNamespace>
-    <AssemblyName>provekit-lsp-csharp</AssemblyName>
+    <RootNamespace>Provekit.Lift.Core</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Provekit.IR\Provekit.IR.csproj" />
-    <ProjectReference Include="..\Provekit.Lift.Core\Provekit.Lift.Core.csproj" />
+    <ProjectReference Include="..\Provekit.Lift.DataAnnotations\Provekit.Lift.DataAnnotations.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.13.0" />
   </ItemGroup>
 
 </Project>
