+++ implementations/go/cmd/provekit-lsp-go/main.go
@@ -23,6 +23,7 @@ import (
 	"strings"
 
 	ir "github.com/tsavo/provekit/go/provekit-ir-symbolic/ir"
+	validator "github.com/tsavo/provekit/go/provekit-lift-go-validator"
 )
 
 type rpcRequest struct {
@@ -194,9 +195,13 @@ func walkSource(src, path string) []ir.Declaration {
 	return decls
 }
 
-// liftStructFromAST walks struct fields with validate tags and lifts to IR.
-// This mirrors the validator.LiftStruct logic but operates on the AST
-// rather than requiring a live struct value.
+// liftStructFromAST walks struct fields with validate tags and lifts to IR
+// by delegating to the shared validator core (task #219).
+//
+// This is the AST-driven counterpart to validator.LiftStruct: rather than
+// requiring a live struct value (reflection), it derives the field's
+// ir.Sort from the AST type expression and calls validator.LiftValidateTags,
+// the same source-agnostic core used by the batch-CLI lift binary.
 func liftStructFromAST(structName string, st *ast.StructType) []ir.Declaration {
 	var decls []ir.Declaration
 
@@ -211,19 +216,19 @@ func liftStructFromAST(structName string, st *ast.StructType) []ir.Declaration {
 		tag = strings.TrimSpace(tag)
 
 		// Parse as a struct tag key:"value"
-		st := reflect.StructTag(tag)
-		validate, ok := st.Lookup("validate")
-		if !ok || validate == "" {
+		structTag := reflect.StructTag(tag)
+		validateTag, ok := structTag.Lookup("validate")
+		if !ok || validateTag == "" {
 			continue
 		}
 
-		// Determine Go type kind from AST
-		goKind := inferGoKind(field.Type)
+		// Derive Sort from AST type expression
+		sort := sortFromASTType(field.Type)
 
 		// Multiple field names? (e.g. `a, b int`)
 		for _, name := range field.Names {
-			// Build IR formula from validate tags
-			f := liftValidateTags(goKind, name.Name, validate)
+			v := ir.MakeVar(name.Name, sort)
+			f := validator.LiftValidateTags(v, sort, validateTag)
 			if f != nil {
 				decls = append(decls, ir.ContractDeclaration{
 					Name:       fmt.Sprintf("%s.%s", structName, name.Name),
@@ -236,145 +241,14 @@ func liftStructFromAST(structName string, st *ast.StructType) []ir.Declaration {
 	return decls
 }
 
-// inferGoKind maps an AST type expression to a rough Go kind string.
-func inferGoKind(expr ast.Expr) string {
-	switch t := expr.(type) {
-	case *ast.Ident:
-		switch t.Name {
-		case "string":
-			return "string"
-		case "int", "int8", "int16", "int32", "int64",
-			"uint", "uint8", "uint16", "uint32", "uint64",
-			"float32", "float64":
-			return "int"
-		case "bool":
-			return "bool"
-		}
-	case *ast.StarExpr, *ast.InterfaceType, *ast.MapType, *ast.ArrayType:
-		return "ref"
+// sortFromASTType reduces an AST type expression to a ProvekIt Sort.
+// Idents are forwarded by name to validator.GoSortFromTypeName; pointers,
+// interfaces, maps, and arrays fall through to ir.Ref.
+func sortFromASTType(expr ast.Expr) ir.Sort {
+	if ident, ok := expr.(*ast.Ident); ok {
+		return validator.GoSortFromTypeName(ident.Name)
 	}
-	return "ref"
-}
-
-// liftValidateTags maps validate tag fragments to IR formulas.
-func liftValidateTags(goKind, varName, tag string) ir.IrFormula {
-	var constraints []ir.IrFormula
-	v := ir.MakeVar(varName, goSortFromKind(goKind))
-
-	parts := strings.Split(tag, ",")
-	for _, part := range parts {
-		part = strings.TrimSpace(part)
-		if part == "" {
-			continue
-		}
-
-		f := liftValidateTag(v, goKind, part)
-		if f != nil {
-			constraints = append(constraints, f)
-		}
-	}
-
-	switch len(constraints) {
-	case 0:
-		return nil
-	case 1:
-		return constraints[0]
-	default:
-		return ir.And(constraints...)
-	}
-}
-
-// liftValidateTag maps a single validate tag to an IR formula.
-// Duplicated from validator.go for AST-mode operation.
-func liftValidateTag(v ir.IrTerm, goKind, tag string) ir.IrFormula {
-	if tag == "required" {
-		if goKind == "string" {
-			return ir.Neq(v, ir.StrConst(""))
-		}
-		return ir.Neq(v, ir.Num(0))
-	}
-
-	for _, op := range []string{"gte=", "lte=", "gt=", "lt=", "eq=", "ne="} {
-		if !strings.HasPrefix(tag, op) {
-			continue
-		}
-		numStr := tag[len(op):]
-		n := 0
-		if _, err := fmt.Sscanf(numStr, "%d", &n); err != nil {
-			return nil
-		}
-		rhs := ir.Num(int64(n))
-		switch op[:len(op)-1] {
-		case "gte":
-			return ir.Gte(v, rhs)
-		case "lte":
-			return ir.Lte(v, rhs)
-		case "gt":
-			return ir.Gt(v, rhs)
-		case "lt":
-			return ir.Lt(v, rhs)
-		case "eq":
-			return ir.Eq(v, rhs)
-		case "ne":
-			return ir.Neq(v, rhs)
-		}
-	}
-
-	if strings.HasPrefix(tag, "min=") {
-		var n int
-		if _, err := fmt.Sscanf(tag[4:], "%d", &n); err != nil {
-			return nil
-		}
-		ni := ir.Num(int64(n))
-		if goKind == "int" {
-			return ir.Gte(v, ni)
-		}
-		return ir.Gte(ir.StringLength(v), ni)
-	}
-	if strings.HasPrefix(tag, "max=") {
-		var n int
-		if _, err := fmt.Sscanf(tag[4:], "%d", &n); err != nil {
-			return nil
-		}
-		ni := ir.Num(int64(n))
-		if goKind == "int" {
-			return ir.Lte(v, ni)
-		}
-		return ir.Lte(ir.StringLength(v), ni)
-	}
-	if strings.HasPrefix(tag, "len=") {
-		var n int
-		if _, err := fmt.Sscanf(tag[4:], "%d", &n); err != nil {
-			return nil
-		}
-		return ir.Eq(ir.StringLength(v), ir.Num(int64(n)))
-	}
-	if tag == "email" {
-		return ir.And() // true placeholder
-	}
-	if strings.HasPrefix(tag, "oneof=") {
-		values := strings.Fields(tag[6:])
-		var eqs []ir.IrFormula
-		for _, val := range values {
-			eqs = append(eqs, ir.Eq(v, ir.StrConst(val)))
-		}
-		return ir.Or(eqs...)
-	}
-	return nil
-}
-
-// goSortFromKind maps a Go kind string to a ProvekIt Sort.
-func goSortFromKind(kind string) ir.Sort {
-	switch kind {
-	case "string":
-		return ir.String
-	case "int":
-		return ir.Int
-	case "bool":
-		return ir.Bool
-	default:
-		return ir.Ref
-	}
+	return ir.Ref
 }
 
 // scanAnnotations scans source lines for //provekit: annotations.
